requirements.txt
pandas 
matplotlib 
requests 
pillow